changes in classes, created new modules files
