+++ Tests/log.html
@@ -1803,13 +1803,13 @@ function stopPropagation(event) {
 window.output = {};
 </script>
 <script type="text/javascript">
-window.output["suite"] = [1,2,3,0,[],[1,0,202],[[4,5,6,0,[],[1,38,162],[],[[7,0,0,[],[1,68,130],[[0,8,0,0,0,0,0,0,[1,69,13],[[0,9,10,0,11,12,13,0,[1,70,8],[[70,2,14],[78,2,15]]],[0,16,10,0,17,0,18,0,[1,78,2],[[80,2,19]]],[0,20,21,0,22,18,0,0,[1,80,1],[]]]],[0,23,0,0,0,0,0,0,[1,83,2],[[0,24,25,0,0,0,26,0,[1,84,0],[[84,2,27]]],[0,28,21,0,29,30,0,0,[1,85,0],[]]]],[0,31,0,0,0,0,0,0,[1,86,2],[[0,32,25,0,0,0,0,0,[1,87,0],[]]]],[0,33,0,0,0,0,0,0,[1,89,3],[[0,34,10,0,35,0,36,0,[1,90,0],[[90,2,37]]],[0,28,21,0,29,38,0,0,[1,90,1],[]]]],[0,39,0,0,0,0,0,0,[1,92,105],[[0,40,10,0,41,42,0,0,[1,93,104],[[93,2,43],[197,2,44]]]]]]]],[],[1,1,0,0]]],[],[],[1,1,0,0]];
+window.output["suite"] = [1,2,3,0,[],[1,0,279],[[4,5,6,0,[],[1,75,201],[],[[7,0,0,[],[1,137,137],[[0,8,0,0,0,0,0,0,[1,139,13],[[0,9,10,0,11,12,13,0,[1,140,9],[[141,2,14],[149,2,15]]],[0,16,10,0,17,0,18,0,[1,149,2],[[150,2,19]]],[0,20,21,0,22,18,0,0,[1,151,1],[]]]],[0,23,0,0,0,0,0,0,[1,153,2],[[0,24,25,0,0,0,26,0,[1,153,1],[[154,2,27]]],[0,28,21,0,29,30,0,0,[1,154,1],[]]]],[0,31,0,0,0,0,0,0,[1,156,2],[[0,32,25,0,0,0,0,0,[1,156,2],[]]]],[0,33,0,0,0,0,0,0,[1,159,1],[[0,34,10,0,35,0,36,0,[1,159,1],[[160,2,37]]],[0,28,21,0,29,38,0,0,[1,160,0],[]]]],[0,39,0,0,0,0,0,0,[1,161,112],[[0,40,10,0,41,42,0,0,[1,162,110],[[162,2,43],[272,2,44]]]]]]]],[],[1,1,0,0]]],[],[],[1,1,0,0]];
 </script>
 <script type="text/javascript">
 window.output["strings"] = [];
 </script>
 <script type="text/javascript">
-window.output["strings"] = window.output["strings"].concat(["*","*Tests","*D:\\OneDrive\\Repositories\\Bookshelf\\Tests","*.","*test Main","*D:\\OneDrive\\Repositories\\Bookshelf\\Tests\\test_Main.robot","*test_Main.robot","*After starting main app as python script, it keeps running in the background","*Start main application using python script","*Start Process","*Process","*<p>Starts a new process on background.\x3c/p>","*${python310}, main.py","*${process}","*Starting process:\nC:\\Python310\\python.exe main.py","*${process} = &lt;Popen: returncode: None args: ['C:\\\\Python310\\\\python.exe', 'main.py']&gt;","*Get Process Id","*<p>Returns the process ID (pid) of the process as an integer.\x3c/p>","*${processid}","*${processid} = 14012","*Log To Console","*BuiltIn","*<p>Logs the given message to the console.\x3c/p>","*Check if database file exists","*Database Exist Check","*settings","*${dataexists}","*${dataexists} = True","*Should Be Equal","*<p>Fails if the given objects are unequal.\x3c/p>","*${dataexists}, ${True}","*Check connection to database","*Connect To Database","*Main application should be running","*Is Process Running","*<p>Checks is the process running or not.\x3c/p>","*${running}","*${running} = True","*${running}, ${True}, msg=APP is not running, but should be","*Close process","*Terminate Process","*<p>Stops the process gracefully or forcefully.\x3c/p>","*kill=True","*Forcefully killing process.","*Process completed."]);
+window.output["strings"] = window.output["strings"].concat(["*","*Tests","*D:\\OneDrive\\Repositories\\Bookshelf\\Tests","*.","*test Main","*D:\\OneDrive\\Repositories\\Bookshelf\\Tests\\test_Main.robot","*test_Main.robot","*After starting main app as python script, it keeps running in the background","*Start main application using python script","*Start Process","*Process","*<p>Starts a new process on background.\x3c/p>","*${python310}, main.py","*${process}","*Starting process:\nC:\\Python310\\python.exe main.py","*${process} = &lt;Popen: returncode: None args: ['C:\\\\Python310\\\\python.exe', 'main.py']&gt;","*Get Process Id","*<p>Returns the process ID (pid) of the process as an integer.\x3c/p>","*${processid}","*${processid} = 7648","*Log To Console","*BuiltIn","*<p>Logs the given message to the console.\x3c/p>","*Check if database file exists","*Database Exist Check","*settings","*${dataexists}","*${dataexists} = True","*Should Be Equal","*<p>Fails if the given objects are unequal.\x3c/p>","*${dataexists}, ${True}","*Check connection to database","*Connect To Database","*Main application should be running","*Is Process Running","*<p>Checks is the process running or not.\x3c/p>","*${running}","*${running} = True","*${running}, ${True}, msg=APP is not running, but should be","*Close process","*Terminate Process","*<p>Stops the process gracefully or forcefully.\x3c/p>","*kill=True","*Forcefully killing process.","*Process completed."]);
 </script>
 <script type="text/javascript">
 window.output["stats"] = [[{"elapsed":"00:00:00","fail":0,"label":"All Tests","pass":1,"skip":0}],[],[{"elapsed":"00:00:00","fail":0,"id":"s1","label":"Tests","name":"Tests","pass":1,"skip":0},{"elapsed":"00:00:00","fail":0,"id":"s1-s1","label":"Tests.test Main","name":"test Main","pass":1,"skip":0}]];
@@ -1818,10 +1818,10 @@ window.output["stats"] = [[{"elapsed":"00:00:00","fail":0,"label":"All Tests","p
 window.output["errors"] = [];
 </script>
 <script type="text/javascript">
-window.output["baseMillis"] = 1668288212756;
+window.output["baseMillis"] = 1668345425053;
 </script>
 <script type="text/javascript">
-window.output["generated"] = 225;
+window.output["generated"] = 319;
 </script>
 <script type="text/javascript">
 window.output["expand_keywords"] = null;
