+++ dbfunctions.py
@@ -1,8 +1,10 @@
 import sqlite3
 from sqlite3 import Error
+import os
 
 
 def create_connection(db_file):
+    """Function to create a connection to a database."""
     conn = None
     try:
         conn = sqlite3.connect(db_file)
@@ -10,3 +12,9 @@ def create_connection(db_file):
         print(e)
 
     return conn
+
+
+def database_exist_check(database_file):
+    """Function to check if a database file exists.
+    Returns True if the database exists, False otherwise."""
+    return os.path.exists(database_file)
