+++ main.py
@@ -1,4 +1,5 @@
-from settings import database_exist_check, connect_to_database
+from dbfunctions import database_exist_check, create_connection
+from settings import *
 from time import sleep
 from ui_mainscreen import Ui_MainWindow
 from PyQt5.QtWidgets import (QApplication, QMainWindow)
@@ -13,15 +14,16 @@ class Window(QMainWindow, Ui_MainWindow):
 
 
 def main():
+    """Functions that creates main application window and establishes connection to database."""
     app = QApplication(sys.argv)
     win = Window()
     win.show()
-    connect_to_database()
+    create_connection(database_file)
     sys.exit(app.exec())
 
 
 if __name__ == '__main__':
-    if database_exist_check():
+    if database_exist_check(database_file):
         main()
     else:
         print("\nCan't find database file!\n")
